feat(theme): Force `color` to black by default, add theming option
better-theming: The CSS selector now also applies `color: black` to go with `background-color: white`.
option-force-light-theme: A new option has been added that forces the colours to always be white for the foreground and black for the background, otherwise inheriting the colours. It is on by default to be as close to original behaviour as possible.